Restore the corrupted database unit test.
